+++ includes/aboutMod.php
@@ -4,30 +4,14 @@
   if(!isset($_GET['aut'])) {
   } else {
     $a['filename'] = 'aboutMod.php';
-    $db = Database::getDB()->getCon();
-    $uName = $db->real_escape_string(stripslashes(trim($_GET['aut'])));
-    $uID = getUserIDbyName($uName);
-    if(!is_int($uID)) {
+    $uName = stripslashes(trim($_GET['aut']));
+
+    $db = Database::getDB();
+    $user = User::newFromName($uName);
+    if(!$user->isLoaded() || !$user->isAdmin()) {
       $a['data']['err'] = $uName;
     } else {
-      $uTxt = '';
-      $sql = 'SELECT
-                About
-              FROM
-                users
-              WHERE
-                ID = ?';
-      $stmt = $db->prepare($sql);
-      if (!$stmt) return $db->error;
-      $stmt->bind_param('i', $uID);
-      if (!$stmt->execute()) return $stmt->error;
-      $stmt->bind_result($uTxt);
-      if(!$stmt->fetch()) return $stmt-error;
-      $stmt->close();
-      $a['data']['about'] = array('id' => $uID,
-                                  'txt' => str_replace('[contactmail]', '</p><address>'.str_replace('@', ' [at] ', getContactMail($db, $uID)).'</address><p>', changetext($uTxt, 'inhalt', $mob)),
-                                  'ClearName' => getClearname($uID),
-                                  'Name' => $uName);
+      $a['data']['about'] = $user;
     }
   }
   return $a;