+++ index.php
@@ -33,7 +33,10 @@
     include('settings/generators.php');
     include('settings/modules.php');
     include('settings/parser.php');
+
+    // load classes
     include('classes/database.php');
+    include('classes/user.php');
  
     $db = Database::getDB()->getCon();
     if($db->connect_errno){